+++ terratex_reallife/ENVIRO/terralapptapp/email_server.lua
@@ -1,45 +1,31 @@
 addEvent("SendeMirMeineEmailsZu", true)
 function SendeMirMeineEmailsZu_func()
-    local query = "SELECT * FROM emails WHERE Empfaenger='" .. getPlayerName(source) .. "'"
-    local result = mysql_query(handler, query)
-    local emailTable = {}
-    if (not result) then
-        outputDebugString("Error executing the query: (" .. mysql_errno(handler) .. ") " .. mysql_error(handler))
-    else
-        local zahler = 0
-        while (mysql_num_rows(result) > zahler) do
-            local dasatz = mysql_fetch_assoc(result)
-            --{Empfänger,Sender,Datum,betreff,Inhalt,new=true,dbid}
-            local readbool = false
-            if (dasatz["neu"] == "1") then
-                readbool = true
-            end
-            table.insert(emailTable, { dasatz["Empfaenger"], dasatz["Sender"], dasatz["Datum"], dasatz["Betreff"], dasatz["Inhalt"], readbool, tonumber(dasatz["ID"]) })
-            --outputChatox(dasatz["Empfaenger"])
-            zahler = zahler + 1
-        end
-    end
-    triggerClientEvent(source, "hierdeineemails", source, emailTable)
-end
+    local emails = MySql.helper.getSync("emails", "*", {Empfaenger = getPlayerName(source)});
+    local emailTable = {};
 
+    for theKey, dasatz in ipairs(emails) do
+        local readbool = false;
+        if (dasatz["neu"] == "1") then
+            readbool = true;
+        end
+        table.insert(emailTable, { dasatz["Empfaenger"], dasatz["Sender"], dasatz["Datum"], dasatz["Betreff"], dasatz["Inhalt"], readbool, tonumber(dasatz["ID"]) });
+    end
+
+    triggerClientEvent(source, "hierdeineemails", source, emailTable);
+end
 addEventHandler("SendeMirMeineEmailsZu", getRootElement(), SendeMirMeineEmailsZu_func)
 
 
 addEvent("setEMailReaded", true)
 function setEMailReaded_func(id)
-    local query = "UPDATE emails SET neu='0' WHERE ID='" .. id .. "'"
-    --outputDebugString(query)
-    mysql_query(handler, query)
+    MySql.helper.update("emails", {neu = 0}, {ID = id});
 end
-
 addEventHandler("setEMailReaded", getRootElement(), setEMailReaded_func)
 
 addEvent("deleteEMail", true)
 function deleteEMail_func(id)
-    local query = "DELETE FROM emails WHERE ID='" .. id .. "'"
-    mysql_query(handler, query)
+    MySql.helper.delete("emails", {ID = id});
 end
-
 addEventHandler("deleteEMail", getRootElement(), deleteEMail_func)
 
 addEvent("sendEMail", true)
@@ -48,10 +34,14 @@ function sendEMail_func(empfaenger, betreff, inhalt)
     if (MySql.helper.existSync("players", { Nickname = empfaenger })) then
 
         if not (vioGetElementData(source, "sendEMailLastSec")) then
-            inhalt = mysql_escape_string(handler, inhalt)
-            betreff = mysql_escape_string(handler, betreff)
-            local query = "INSERT INTO emails (Empfaenger,Sender,Betreff,Inhalt) VALUES ('" .. empfaenger .. "','" .. sender .. "','" .. betreff .. "','" .. inhalt .. "');"
-            mysql_query(handler, query)
+
+            MySql.helper.insert("emails", {
+                Empfaenger = empfaenger,
+                Sender = sender,
+                Betreff = betreff,
+                Inhalt = inhalt
+            });
+
             outputChatBox("EMail wurde versendet!", source, 0, 255, 0)
             vioSetElementData(source, "sendEMailLastSec", true)
             setTimer(vioSetElementData, 5000, 1, source, "sendEMailLastSec", false)
@@ -62,13 +52,4 @@ function sendEMail_func(empfaenger, betreff, inhalt)
         outputChatBox("Dieser Empfänger existiert nicht!", source, 255, 0, 0)
     end
 end
-
 addEventHandler("sendEMail", getRootElement(), sendEMail_func)
-
-
-
-
-
-
-
-